use one and the same  sha256 function

+++ tools/common/logging_wrapper.py
@@ -2,7 +2,7 @@ import logging
 import os
 
 
-def setup_logging(logger_name=None, log_file_name=None):
+def setup_logging(logger_name=None, log_file_name=None, append_mode=False):
     if logger_name is None:
         if log_file_name is not None:
             logger_name,_ = os.path.splitext(log_file_name)
@@ -19,9 +19,12 @@ def setup_logging(logger_name=None, log_file_name=None):
 
     # create file handler which logs even debug messages
     if log_file_name is not None:
-        if os.path.exists(log_file_name):
-            os.remove(log_file_name)
-        fh = logging.FileHandler(log_file_name, encoding="utf8")
+        if append_mode:
+            fh = logging.FileHandler(logfilename, "a+", encoding="utf8")
+        else:
+            if os.path.exists(log_file_name):
+                os.remove(log_file_name)
+            fh = logging.FileHandler(log_file_name, encoding="utf8")
         fh.setLevel(logging.DEBUG)
         fh.setFormatter(formatter)
         logger.addHandler(fh)
